+++ src/lib/linear-scale.js
@@ -1,0 +1,10 @@
+export default function linearScale(domain, range) {
+  return function(value) {
+    if (domain[0] === domain[1] || range[0] === range[1]) {
+      return range[0];
+    }
+    var ratio = (range[1] - range[0]) / (domain[1] - domain[0]),
+      result = range[0] + ratio * (value - domain[0]);
+    return result;
+  };
+}